requirements.txt
python-dotenv==1.0.0
python-telegram-bot~=13.0
vk_api==11.9.9
google-cloud-dialogflow==2.19.1

